skills/tech-digest/requirements.txt
feedparser>=6.0
requests>=2.28
scikit-learn>=1.3
pyyaml>=6.0
pytest>=7.0
